call reward from raw substrate api
